Refactor MockProperties and use in testSubmitQueuedCart
